requirements.txt
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
fire==0.7.0
idna==3.10
isort==6.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
python-dotenv==1.1.0
requests==2.32.4
termcolor==3.1.0
urllib3==2.4.0
